altered comment delete permissions to admins, curators and those that can edit the object that the comment relates to
git-svn-id: svn+ssh://rubyforge.org/var/svn/myexperiment/trunk@2307 4f7b4649-9ca9-420a-81b3-55a188e4d46f

+++ app/controllers/workflows_controller.rb
@@ -90,10 +90,7 @@ class WorkflowsController < ApplicationController
   def comment_delete
     if params[:comment_id]
       comment = Comment.find(params[:comment_id].to_i)
-      # security checks:
-      if comment.user_id == current_user.id and comment.commentable_type.downcase == 'workflow' and comment.commentable_id == @workflow.id
-        comment.destroy
-      end
+      comment.destroy if Authorization.check(:action => 'destroy', :object => comment, :user => current_user)
     end
     
     respond_to do |format|
